@@ -1,6 +1,6 @@
 package com.flipcam.model;
 
-import android.support.annotation.NonNull;
+import androidx.annotation.NonNull;
 
 public class Dimension implements Comparable<Dimension>{
 
