@@ -24,8 +24,6 @@ import android.os.Bundle;
 import android.os.Environment;
 import android.os.StatFs;
 import android.preference.PreferenceManager;
-import android.support.annotation.Nullable;
-import android.support.v4.app.Fragment;
 import android.util.Log;
 import android.util.Range;
 import android.view.Gravity;
@@ -42,6 +40,9 @@ import android.widget.LinearLayout;
 import android.widget.SeekBar;
 import android.widget.TextView;
 import android.widget.Toast;
+
+import androidx.annotation.Nullable;
+import androidx.fragment.app.Fragment;
 
 import com.flipcam.constants.Constants;
 import com.flipcam.data.MediaTableConstants;
@@ -183,7 +184,7 @@ public class VideoFragment extends Fragment{
     }
 
     @Override
-    public void onCreate(@androidx.annotation.Nullable @Nullable Bundle savedInstanceState) {
+    public void onCreate(@androidx.annotation.Nullable Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         setRetainInstance(true);
     }
@@ -1183,6 +1184,9 @@ public class VideoFragment extends Fragment{
                 getResources().getString(R.string.phoneLocation) : getResources().getString(R.string.sdcardLocation);
         mediaLocEdit.putString(Constants.MEDIA_LOCATION_VIEW_SELECT, mediaLocValue);
         mediaLocEdit.commit();
+        if(controlVisbilityPreference == null){
+            controlVisbilityPreference = (ControlVisbilityPreference)getApplicationContext();
+        }
         controlVisbilityPreference.setFromGallery(false);
         startActivity(mediaIntent);
     }
